requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
